refactor: styled -> dynamic/static, animation
