+++ src/components/UI/Adsense/Adsense.tsx
@@ -3,7 +3,7 @@ import { useEffect, useState } from 'react';
 
 import { GOOGLE_ADSENSE } from '@/constant';
 import { isMobile } from '@/ui/lib/mediaQuery';
-import { styled } from '@/ui/styled';
+import { styled } from '@/ui/styled/dynamic';
 
 type Size = {
   adsWidth?: number;
