+++ src/components/UI/Layout/Cluster.tsx
@@ -1,8 +1,8 @@
 import type { AriaRole, ReactNode } from 'react';
 import { memo, useMemo } from 'react';
 
-import { css, styled } from '@/ui/styled';
 import type { SpaceGap } from '@/ui/styled/CssBaseline/Settings/Space';
+import { css, styled } from '@/ui/styled/dynamic';
 
 type Props = {
   as?: keyof JSX.IntrinsicElements;
