+++ src/components/UI/Search/SearchHeader.tsx
@@ -1,7 +1,7 @@
 import { useId, useRef } from 'react';
 
 import { ICON_SIZE_XS, MagnifyingGlassIcon } from '@/ui/icons';
-import { styled } from '@/ui/styled';
+import { styled } from '@/ui/styled/dynamic';
 
 import type { onKeyupProps } from './type';
