+++ src/components/UI/Search/SearchPanel.tsx
@@ -3,9 +3,8 @@ import { memo, useMemo } from 'react';
 import { Anchor as _Anchor } from '@/components/UI/Anchor';
 import { useRouteChangeComplete } from '@/hooks/useRouteChangeComplete';
 import escapeHTML from '@/lib/escapeHTML';
-import { fadeIn } from '@/ui/animation';
 import { isMobile } from '@/ui/lib/mediaQuery';
-import { styled } from '@/ui/styled';
+import { styled } from '@/ui/styled/dynamic';
 
 import type { SearchProps, onCloseDialogProps } from './type';
 import { useSearchHeader } from './useSearchHeader';
@@ -89,7 +88,7 @@ const SearchMain = styled.div`
   border-radius: var(--border-radius-4);
   box-shadow: 0 3px 6px rgba(0, 0, 0, 0.16);
   opacity: 0;
-  animation: ${fadeIn} 0.8s ease;
+  animation: fadeIn 0.8s ease;
   animation-fill-mode: both;
 
   ${isMobile} {
