+++ src/components/UI/Tag/index.tsx
@@ -2,7 +2,7 @@ import { memo } from 'react';
 
 import { Anchor } from '@/components/UI/Anchor';
 import { TAG_VIEW_LIMIT } from '@/constant';
-import { css, styled } from '@/ui/styled';
+import { styled } from '@/ui/styled/dynamic';
 
 export type Props = {
   slug: string;
@@ -31,6 +31,7 @@ const PostTag = memo(function PostTag({ tags, hasRelTag = true }: PostTagProps) 
 
       return isAnchor ? (
         <TagAnchor
+          className="post-tag-anchor"
           href={`/tags/${slug}`}
           key={slug}
           {...(hasRelTag && {
@@ -41,7 +42,7 @@ const PostTag = memo(function PostTag({ tags, hasRelTag = true }: PostTagProps) 
           <Count aria-hidden="true">{count}</Count>
         </TagAnchor>
       ) : (
-        <DisabledTagAnchor aria-hidden="true" key={slug}>
+        <DisabledTagAnchor className="post-tag-anchor" aria-hidden="true" key={slug}>
           {slug}
         </DisabledTagAnchor>
       );
@@ -49,19 +50,6 @@ const PostTag = memo(function PostTag({ tags, hasRelTag = true }: PostTagProps) 
 });
 
 export default PostTag;
-
-export const PostTagAnchorStyle = css`
-  display: inline-flex;
-  justify-content: center;
-  padding: var(--space-½) var(--space-2);
-  font-size: var(--font-size-sm);
-  line-height: var(--line-height-md);
-  color: var(--color-gray-11);
-  text-align: center;
-  white-space: nowrap;
-  background-color: var(--color-gray-3A);
-  border-radius: var(--border-radius-8);
-`;
 
 const Count = styled.span`
   display: inline-flex;
@@ -79,8 +67,6 @@ const Count = styled.span`
 `;
 
 const TagAnchor = styled(Anchor)`
-  ${PostTagAnchorStyle}
-
   &:hover {
     background-color: var(--color-gray-4A);
   }
@@ -91,8 +77,6 @@ const TagAnchor = styled(Anchor)`
 `;
 
 const DisabledTagAnchor = styled.span`
-  ${PostTagAnchorStyle}
-
   &:hover {
     cursor: not-allowed;
   }
