+++ src/components/UI/Toast/Toast.tsx
@@ -1,8 +1,7 @@
 import type { ForwardedRef } from 'react';
 import { forwardRef, useMemo } from 'react';
 
-import { fadeIn } from '@/ui/animation';
-import { styled } from '@/ui/styled';
+import { styled } from '@/ui/styled/dynamic';
 
 import { useToast } from './useToast';
 
@@ -53,6 +52,6 @@ const Container = styled.div`
   opacity: 0;
 
   &[aria-hidden='true'] {
-    animation: ${fadeIn} 0.4s linear both;
+    animation: fadeIn 0.4s linear both;
   }
 `;
