+++ src/pages/tags/[slug]/_components/Pagination.tsx
@@ -3,10 +3,9 @@ import type { ReactNode } from 'react';
 import { memo, useCallback, useEffect, useMemo, useState } from 'react';
 
 import { Cluster, Stack } from '@/components/UI/Layout';
-import { easeOut } from '@/ui/foundation/easing';
 import { CaretLeftIcon, CaretRightIcon, ICON_SIZE_LG, ICON_SIZE_XS } from '@/ui/icons';
 import { isDesktop, isMobile } from '@/ui/lib/mediaQuery';
-import { css, styled } from '@/ui/styled';
+import { css, styled } from '@/ui/styled/dynamic';
 
 type Props = {
   displayCount: number;
@@ -125,7 +124,7 @@ const PaginationButtonStyle = css`
   color: var(--color-gray-11);
   cursor: pointer;
   border-radius: var(--border-radius-full);
-  transition: background-color 0.1s ${easeOut};
+  transition: background-color 0.1s var(--easing-ease-out);
 
   &[disabled] {
     cursor: not-allowed;
